Revise code conventions in string utility
This update would remove the infinite loop, replacing it with a simpler loop with consistent flow of control. It would also make the code more portable and executable by adding the shebang. One should also consider replacing the sentinel/exit token from "start" to a different one, like the empty string, but this is the lead developer's call.